+++ test/test_matrix_print.cpp
@@ -7,7 +7,7 @@ int TestMatrixPrint() {
     AMatrix::Matrix<double, 3, 3> a_matrix;
 
     for (size_t i = 0; i < 9; i++) {
-        a_matrix[i] = i * i;
+        a_matrix[i] = static_cast<double>(i * i);
     }
 
     std::stringstream string_stream;
